% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaCorrelationMatrix.R
\name{apa.cor.matrix.default}
\alias{apa.cor.matrix.default}
\title{Default method to generate a correlation matrix with values}
\usage{
\method{apa.cor.matrix}{default}(data = data.frame(), position = "lower", sig = TRUE)
}
\arguments{
\item{data}{Raw dataset with variables.}

\item{position}{(optional) Specify whether the correlations should be displayed in the \code{upper}, or \code{lower} diagonal of the table.}

\item{sig}{(optional) Specify whether the significance should be displayed in a separate column.}
}
\value{
\code{apa.cor.matrix} object; a list consisting of
\item{succes}{message in case of an error}
\item{data}{the data with correlation values}
\item{smallest}{the smallest \emph{r} value which is significant at \emph{p} < .05.}
}
\description{
Default method to generate a correlation matrix with values
}
\examples{

# Use apa.cor.matrix function
apa.cor.matrix(
  data = data.frame(
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1)
  ),
  position = "upper"
)
}
