% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apaCorrelationMatrix.R
\name{apa.cor.matrix}
\alias{apa.cor.matrix}
\title{Generic method to generate a correlation matrix with values}
\usage{
apa.cor.matrix(data = data.frame(), position = "lower")
}
\arguments{
\item{data}{Raw dataset with variables.}

\item{position}{(optional) Specify whether the correlations should be displayed in the \code{upper}, or \code{lower} diagonal of the table.}
}
\value{
\code{apa.cor.matrix} object; a list consisting of
\item{succes}{message in case of an error}
\item{data}{the data with correlation values}
}
\description{
Generic method to generate a correlation matrix with values
}
\examples{
# Use apa.cor.matrix function
apa.cor.matrix(
  data = data.frame(
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1),
    rnorm(100, mean = 0, sd = 1)
  ),
  position = "upper"
)
}

