% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaMergeData.R
\name{apa.merge}
\alias{apa.merge}
\title{Generic method to merge two vectors and create a header}
\usage{
apa.merge(a = NULL, b = NULL, header = NULL)
}
\arguments{
\item{a}{Values in the first column.}

\item{b}{Values in the second column.}

\item{header}{Vector with both column names.}
}
\value{
\code{apa.merge} object; a list consisting of
\item{succes}{message in case of an error}
\item{data}{the merged data}
\item{header}{the header for the merged data}
}
\description{
Generic method to merge two vectors and create a header
}
\examples{

# Use apa.merge function
apa.merge(a = rnorm(100, mean = 0, sd = 1), b = rnorm(100, mean = 0, sd = 1), header = c("M", "SD"))
}

