% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaRegression.R
\name{apa.regression}
\alias{apa.regression}
\title{Generic method to generate an APA style table with regression parameters for MS Word.}
\usage{
apa.regression(..., variables = NULL, number = "XX", title = "APA Table",
  filename = "APA Table.docx", note = NULL, landscape = FALSE,
  save = TRUE, type = "wide")
}
\arguments{
\item{...}{Regression (i.e., lm) result objects.}

\item{variables}{The variable names for in the table.}

\item{number}{(optional) The table number in the document.}

\item{title}{(optional) Name of the table.}

\item{filename}{(optional) Specify the filename (including valid '\code{.docx}' extension).}

\item{note}{(optional) Add a footnote to the bottom of the table.}

\item{landscape}{(optional) Set (\code{TRUE}) if the table should be generated in landscape mode.}

\item{save}{(optional) Set (\code{FALSE}) if the table should not be saved in a document.}

\item{type}{(optional) Not implemented.}
}
\value{
\code{apa.regression} object; a list consisting of
\item{succes}{message in case of an error}
\item{save}{flag which indicates whther the document is saved}
\item{data}{dataset with regression statistics}
\item{table}{\code{FlexTable {ReporteRs}} object}
}
\description{
Generic method to generate an APA style table with regression parameters for MS Word.
}

