% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apaSignificance.R
\name{apa.signif}
\alias{apa.signif}
\title{Generic method to make a footnote indicating significant values.}
\usage{
apa.signif(data = data.frame())
}
\arguments{
\item{data}{Dataset with statistics.}
}
\value{
\code{apa.signif} object; a list consisting of
\item{succes}{message in case of an error}
\item{signif}{\code{pot {ReporteRs}} object}
}
\description{
Generic method to make a footnote indicating significant values.
}
\examples{
# Specify statistics
example <- data.frame(
  c("Column 1", "Column 2", "Column 3"),
  c(3.45, 5.21, 2.64),
  c("**", "", "***")
)

# Use apa.descriptives function
apa.signif(data = example)
}

