% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaCorrelationMatrix.R
\name{print.apa.cor.matrix}
\alias{print.apa.cor.matrix}
\title{Define a print method}
\usage{
\method{print}{apa.cor.matrix}(x, ...)
}
\arguments{
\item{x}{A \code{apa.cor.matrix} object}
}
\description{
Define a print method
}
