% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaDescriptives.R
\name{print.apa.descriptives}
\alias{print.apa.descriptives}
\title{Define a print method}
\usage{
\method{print}{apa.descriptives}(x, ...)
}
\arguments{
\item{x}{A \code{apa.descriptives} object}
}
\description{
Define a print method
}
