% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaMergeData.R
\name{print.apa.merge}
\alias{print.apa.merge}
\title{Define a print method}
\usage{
\method{print}{apa.merge}(x, ...)
}
\arguments{
\item{x}{A \code{apa.merge} object}
}
\description{
Define a print method
}
