% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaRegression.R
\name{print.apa.regression}
\alias{print.apa.regression}
\title{Define a print method}
\usage{
\method{print}{apa.regression}(x, ...)
}
\arguments{
\item{x}{A \code{apa.regression} object}
}
\description{
Define a print method
}
