% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaSignificance.R
\name{print.apa.signif}
\alias{print.apa.signif}
\title{Define a print method}
\usage{
\method{print}{apa.signif}(x, ...)
}
\arguments{
\item{x}{A \code{apa.signif} object}
}
\description{
Define a print method
}
